Document additional step after restore
In some usecases (e.g. when the restore is performed on the same host where the backup was executed) the web UI shows wrong data, because of the OTOBO cache.

To solve this a new step to clear the OTOBO cache was documented, which should be performed after successful restore.